moving test to make windows test pass
